Implement `ramp_a` and move test helpers to test_utils (#44)
* Add admin_key to state

* Implement ramp_a

* Move test functions to test_utils

* Update comments

* Update imports

* Update client

* Actually implement ramp_a

* Use MAX_MAP and MIN_AMP constants

* Add ramp_a instruction and test helper

* Add ramp_a tests

* Use unchecked mult (should never overflow here)
