+++ src/admin.rs
@@ -3,19 +3,22 @@
 #![cfg(feature = "program")]
 
 use crate::{
+    bn::U256,
+    curve::{StableSwap, MAX_AMP, MIN_AMP, MIN_RAMP_DURATION},
     error::SwapError,
     fees::Fees,
     instruction::{AdminInstruction, RampAData},
+    state::SwapInfo,
+    utils::authority_id,
 };
-#[cfg(not(target_arch = "bpf"))]
 use solana_sdk::{
-    account_info::AccountInfo, entrypoint::ProgramResult, info, program_error::ProgramError,
+    account_info::{next_account_info, AccountInfo},
+    entrypoint::ProgramResult,
+    info,
+    program_error::ProgramError,
+    program_pack::Pack,
     pubkey::Pubkey,
-};
-#[cfg(target_arch = "bpf")]
-use solana_sdk::{
-    account_info::AccountInfo, entrypoint::ProgramResult, info, program_error::ProgramError,
-    pubkey::Pubkey,
+    sysvar::{clock::Clock, Sysvar},
 };
 
 /// Process admin instruction
@@ -26,11 +29,11 @@ pub fn process_admin_instruction(
 ) -> ProgramResult {
     match *instruction {
         AdminInstruction::RampA(RampAData {
-            future_amp,
+            target_amp,
             stop_ramp_ts,
         }) => {
             info!("Instruction : RampA");
-            ramp_a(program_id, future_amp, stop_ramp_ts, accounts)
+            ramp_a(program_id, target_amp, stop_ramp_ts, accounts)
         }
         AdminInstruction::StopRampA => {
             info!("Instruction: StopRampA");
@@ -80,7 +83,7 @@ pub fn process_admin_instruction(
 }
 
 /// Access control for admin only instructions
-fn _is_admin(expected_admin_key: &Pubkey, admin_account_info: &AccountInfo) -> ProgramResult {
+fn is_admin(expected_admin_key: &Pubkey, admin_account_info: &AccountInfo) -> ProgramResult {
     if expected_admin_key != admin_account_info.key {
         return Err(SwapError::Unauthorized.into());
     }
@@ -92,12 +95,73 @@ fn _is_admin(expected_admin_key: &Pubkey, admin_account_info: &AccountInfo) -> P
 
 /// Ramp to future a
 fn ramp_a(
-    _program_id: &Pubkey,
-    _future_a: u64,
-    _stop_ramp_ts: i64,
-    _account: &[AccountInfo],
+    program_id: &Pubkey,
+    target_amp: u64,
+    stop_ramp_ts: i64,
+    accounts: &[AccountInfo],
 ) -> ProgramResult {
-    unimplemented!("ramp_a not implemented");
+    let account_info_iter = &mut accounts.iter();
+    let swap_info = next_account_info(account_info_iter)?;
+    let authority_info = next_account_info(account_info_iter)?;
+    let admin_info = next_account_info(account_info_iter)?;
+    let clock_sysvar_info = next_account_info(account_info_iter)?;
+
+    if target_amp < MIN_AMP || target_amp > MAX_AMP {
+        return Err(SwapError::InvalidInput.into());
+    }
+    let mut token_swap = SwapInfo::unpack(&swap_info.data.borrow())?;
+    is_admin(&token_swap.admin_key, admin_info)?;
+    if *authority_info.key != authority_id(program_id, swap_info.key, token_swap.nonce)? {
+        return Err(SwapError::InvalidProgramAddress.into());
+    }
+
+    let clock = Clock::from_account_info(clock_sysvar_info)?;
+    let ramp_lock_ts = token_swap
+        .start_ramp_ts
+        .checked_add(MIN_RAMP_DURATION)
+        .ok_or(SwapError::CalculationFailure)?;
+    if clock.unix_timestamp < ramp_lock_ts {
+        return Err(SwapError::RampLocked.into());
+    }
+    let min_ramp_ts = clock
+        .unix_timestamp
+        .checked_add(MIN_RAMP_DURATION)
+        .ok_or(SwapError::CalculationFailure)?;
+    if stop_ramp_ts < min_ramp_ts {
+        return Err(SwapError::InsufficientRampTime.into());
+    }
+
+    const MAX_A_CHANGE: u64 = 10;
+    let invariant = StableSwap::new(
+        token_swap.initial_amp_factor,
+        token_swap.target_amp_factor,
+        clock.unix_timestamp,
+        token_swap.start_ramp_ts,
+        token_swap.stop_ramp_ts,
+    );
+    let current_amp = U256::to_u64(
+        invariant
+            .compute_amp_factor()
+            .ok_or(SwapError::CalculationFailure)?,
+    )?;
+    if target_amp < current_amp {
+        if current_amp > target_amp * MAX_A_CHANGE {
+            // target_amp too low
+            return Err(SwapError::InvalidInput.into());
+        }
+    } else {
+        if target_amp > current_amp * MAX_A_CHANGE {
+            // target_amp too high
+            return Err(SwapError::InvalidInput.into());
+        }
+    }
+
+    token_swap.initial_amp_factor = current_amp;
+    token_swap.target_amp_factor = target_amp;
+    token_swap.start_ramp_ts = clock.unix_timestamp;
+    token_swap.stop_ramp_ts = stop_ramp_ts;
+    SwapInfo::pack(token_swap, &mut swap_info.data.borrow_mut())?;
+    Ok(())
 }
 
 /// Stop ramp a
@@ -162,8 +226,14 @@ fn revert_new_fees(_program_id: &Pubkey, _accounts: &[AccountInfo]) -> ProgramRe
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utils::test_utils::pubkey_rand;
+    use crate::{
+        curve::ZERO_TS,
+        utils::{test_utils::*, TOKEN_PROGRAM_ID},
+    };
     use solana_sdk::clock::Epoch;
+
+    const DEFAULT_TOKEN_A_AMOUNT: u64 = 1_000_000_000;
+    const DEFAULT_TOKEN_B_AMOUNT: u64 = 1_000_000_000;
 
     #[test]
     fn test_is_admin() {
@@ -183,7 +253,7 @@ mod tests {
         );
 
         // Correct admin
-        assert_eq!(Ok(()), _is_admin(&admin_key, &admin_account_info));
+        assert_eq!(Ok(()), is_admin(&admin_key, &admin_account_info));
 
         // Unauthorized account
         let fake_admin_key = pubkey_rand();
@@ -191,14 +261,125 @@ mod tests {
         fake_admin_account.key = &fake_admin_key;
         assert_eq!(
             Err(SwapError::Unauthorized.into()),
-            _is_admin(&admin_key, &fake_admin_account)
+            is_admin(&admin_key, &fake_admin_account)
         );
 
         // Admin did not sign
         admin_account_info.is_signer = false;
         assert_eq!(
             Err(ProgramError::MissingRequiredSignature),
-            _is_admin(&admin_key, &admin_account_info)
+            is_admin(&admin_key, &admin_account_info)
         );
+    }
+
+    #[test]
+    fn test_ramp_a() {
+        let user_key = pubkey_rand();
+        let amp_factor = MIN_AMP * 100;
+        let mut accounts = SwapAccountInfo::new(
+            &user_key,
+            amp_factor,
+            DEFAULT_TOKEN_A_AMOUNT,
+            DEFAULT_TOKEN_B_AMOUNT,
+            DEFAULT_TEST_FEES,
+        );
+
+        // swap not initialized
+        {
+            assert_eq!(
+                Err(ProgramError::UninitializedAccount),
+                accounts.ramp_a(MIN_AMP, ZERO_TS, MIN_RAMP_DURATION)
+            );
+        }
+
+        accounts.initialize_swap().unwrap();
+
+        // Invalid target amp
+        {
+            let stop_ramp_ts = MIN_RAMP_DURATION;
+            let target_amp = 0;
+            assert_eq!(
+                Err(SwapError::InvalidInput.into()),
+                accounts.ramp_a(target_amp, ZERO_TS, stop_ramp_ts)
+            );
+            let target_amp = MAX_AMP + 1;
+            assert_eq!(
+                Err(SwapError::InvalidInput.into()),
+                accounts.ramp_a(target_amp, ZERO_TS, stop_ramp_ts)
+            );
+        }
+
+        // wrong nonce for authority_key
+        {
+            let old_authority = accounts.authority_key;
+            let (bad_authority_key, _nonce) = Pubkey::find_program_address(
+                &[&accounts.swap_key.to_bytes()[..]],
+                &TOKEN_PROGRAM_ID,
+            );
+            accounts.authority_key = bad_authority_key;
+            assert_eq!(
+                Err(SwapError::InvalidProgramAddress.into()),
+                accounts.ramp_a(MIN_AMP, ZERO_TS, MIN_RAMP_DURATION)
+            );
+            accounts.authority_key = old_authority;
+        }
+
+        // Unauthorized account
+        {
+            let old_admin_key = accounts.admin_key;
+            let fake_admin_key = pubkey_rand();
+            accounts.admin_key = fake_admin_key;
+            assert_eq!(
+                Err(SwapError::Unauthorized.into()),
+                accounts.ramp_a(MIN_AMP, ZERO_TS, MIN_RAMP_DURATION)
+            );
+            accounts.admin_key = old_admin_key;
+        }
+
+        // ramp locked
+        {
+            assert_eq!(
+                Err(SwapError::RampLocked.into()),
+                accounts.ramp_a(MIN_AMP, ZERO_TS, MIN_RAMP_DURATION / 2)
+            );
+        }
+
+        // insufficient ramp time
+        {
+            assert_eq!(
+                Err(SwapError::InsufficientRampTime.into()),
+                accounts.ramp_a(amp_factor, MIN_RAMP_DURATION, ZERO_TS)
+            );
+        }
+
+        // invalid amp targets
+        {
+            // amp target too low
+            assert_eq!(
+                Err(SwapError::InvalidInput.into()),
+                accounts.ramp_a(MIN_AMP, MIN_RAMP_DURATION, MIN_RAMP_DURATION * 2)
+            );
+            // amp target too high
+            assert_eq!(
+                Err(SwapError::InvalidInput.into()),
+                accounts.ramp_a(MAX_AMP, MIN_RAMP_DURATION, MIN_RAMP_DURATION * 2)
+            );
+        }
+
+        // valid ramp
+        {
+            let target_amp = MIN_AMP * 200;
+            let current_ts = MIN_RAMP_DURATION;
+            let stop_ramp_ts = MIN_RAMP_DURATION * 2;
+            accounts
+                .ramp_a(target_amp, current_ts, stop_ramp_ts)
+                .unwrap();
+
+            let swap_info = SwapInfo::unpack(&accounts.swap_account.data).unwrap();
+            assert_eq!(swap_info.initial_amp_factor, amp_factor);
+            assert_eq!(swap_info.target_amp_factor, target_amp);
+            assert_eq!(swap_info.start_ramp_ts, current_ts);
+            assert_eq!(swap_info.stop_ramp_ts, stop_ramp_ts);
+        }
     }
 }
